Update and cleanups

+++ src/main/java/gg/dystellar/core/api/comms/Channel.java
@@ -220,7 +220,7 @@ public final class Channel {
 		}
 
 		public void writeByte(byte b) {
-			this.buf.put((byte)b);
+			this.buf.put(b);
 		}
 
 		@Override
